Applied fixes from trunk to ReSync branch, fixed up additional conflicts, errors, and warnings.

+++ GPL/branches/alsa-resync1/alsa-kernel/core/seq/seq_clientmgr.c
@@ -2180,7 +2180,7 @@ static int snd_seq_do_ioctl(struct snd_seq_client *client, unsigned int cmd,
 }
 
 
-static long snd_seq_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
+static int snd_seq_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
 	struct snd_seq_client *client = file->private_data;
